refactor: 서버에 답변을 제출하는 기능을  useSubmitAnswer 훅으로 묶음
